stable discource

+++ app/services/push_notification_pusher.rb
@@ -2,12 +2,11 @@
 
 class PushNotificationPusher
   TOKEN_VALID_FOR_SECONDS ||= 5 * 60
-  CONNECTION_TIMEOUT_SECONDS = 5
 
   def self.push(user, payload)
     I18n.with_locale(user.effective_locale) do
       message = {
-        title: payload[:translated_title] || I18n.t(
+        title: I18n.t(
           "discourse_push_notifications.popup.#{Notification.types[payload[:notification_type]]}",
           site_title: SiteSetting.title,
           topic: payload[:topic_title],
@@ -23,6 +22,7 @@ class PushNotificationPusher
       }
 
       subscriptions(user).each do |subscription|
+        subscription = JSON.parse(subscription.data)
         send_notification(user, subscription, message)
       end
     end
@@ -36,12 +36,12 @@ class PushNotificationPusher
     user.push_subscriptions.clear
   end
 
-  def self.subscribe(user, push_params, send_confirmation)
-    data = push_params.to_json
+  def self.subscribe(user, subscription, send_confirmation)
+    data = subscription.to_json
     subscriptions = PushSubscription.where(user: user, data: data)
     subscriptions_count = subscriptions.count
 
-    new_subscription = if subscriptions_count > 1
+    if subscriptions_count > 1
       subscriptions.destroy_all
       PushSubscription.create!(user: user, data: data)
     elsif subscriptions_count == 0
@@ -58,13 +58,15 @@ class PushNotificationPusher
         tag: "#{Discourse.current_hostname}-subscription"
       }
 
-      send_notification(user, new_subscription, message)
+      send_notification(user, subscription, message)
     end
   end
 
   def self.unsubscribe(user, subscription)
     PushSubscription.find_by(user: user, data: subscription.to_json)&.destroy!
   end
+
+  protected
 
   def self.get_badge
     if (url = SiteSetting.site_push_notifications_icon_url).present?
@@ -74,40 +76,13 @@ class PushNotificationPusher
     end
   end
 
-  MAX_ERRORS ||= 3
-  MIN_ERROR_DURATION ||= 86400 # 1 day
-
-  def self.handle_generic_error(subscription, error, user, endpoint, message)
-    subscription.error_count += 1
-    subscription.first_error_at ||= Time.zone.now
-
-    delta = Time.zone.now - subscription.first_error_at
-    if subscription.error_count >= MAX_ERRORS && delta > MIN_ERROR_DURATION
-      subscription.destroy!
-    else
-      subscription.save!
-    end
-
-    Discourse.warn_exception(
-      error,
-      message: "Failed to send push notification",
-      env: {
-        user_id: user.id,
-        endpoint: endpoint,
-        message: message.to_json
-      }
-    )
-  end
-
   def self.send_notification(user, subscription, message)
-    parsed_data = subscription.parsed_data
-
-    endpoint = parsed_data["endpoint"]
-    p256dh = parsed_data.dig("keys", "p256dh")
-    auth = parsed_data.dig("keys", "auth")
+    endpoint = subscription["endpoint"]
+    p256dh = subscription.dig("keys", "p256dh")
+    auth = subscription.dig("keys", "auth")
 
     if (endpoint.blank? || p256dh.blank? || auth.blank?)
-      subscription.destroy!
+      unsubscribe(user, subscription)
       return
     end
 
@@ -122,29 +97,24 @@ class PushNotificationPusher
           public_key: SiteSetting.vapid_public_key,
           private_key: SiteSetting.vapid_private_key,
           expiration: TOKEN_VALID_FOR_SECONDS
-        },
-        open_timeout: CONNECTION_TIMEOUT_SECONDS,
-        read_timeout: CONNECTION_TIMEOUT_SECONDS,
-        ssl_timeout: CONNECTION_TIMEOUT_SECONDS
+        }
       )
-
-      if subscription.first_error_at || subscription.error_count != 0
-        subscription.update_columns(error_count: 0, first_error_at: nil)
-      end
     rescue Webpush::ExpiredSubscription
-      subscription.destroy!
+      unsubscribe(user, subscription)
     rescue Webpush::ResponseError => e
       if e.response.message == "MismatchSenderId"
-        subscription.destroy!
+        unsubscribe(user, subscription)
       else
-        handle_generic_error(subscription, e, user, endpoint, message)
+        Discourse.warn_exception(
+          e,
+          message: "Failed to send push notification",
+          env: {
+            user_id: user.id,
+            endpoint: subscription["endpoint"],
+            message: message.to_json
+          }
+        )
       end
-    rescue Timeout::Error => e
-      handle_generic_error(subscription, e, user, endpoint, message)
     end
   end
-
-  private_class_method :send_notification
-  private_class_method :handle_generic_error
-
 end
